Renamed and removed CompilerLocatorMSVC files.
