+++ Plugin/CompilerLocatorMSVC.cpp
@@ -23,7 +23,7 @@
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
 
-#include "CompilerLocatorMSVCBase.h"
+#include "CompilerLocatorMSVC.h"
 #include "compiler.h"
 #include <globals.h>
 #include <wx/tokenzr.h>
