--- Plugin/CompilerLocatorMSVC10.h
@@ -1,42 +1,0 @@
-//////////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////////
-//
-// copyright            : (C) 2014 Eran Ifrah
-// file name            : CompilerLocatorMSVC10.h
-//
-// -------------------------------------------------------------------------
-// A
-//              _____           _      _     _ _
-//             /  __ \         | |    | |   (_) |
-//             | /  \/ ___   __| | ___| |    _| |_ ___
-//             | |    / _ \ / _  |/ _ \ |   | | __/ _ )
-//             | \__/\ (_) | (_| |  __/ |___| | ||  __/
-//              \____/\___/ \__,_|\___\_____/_|\__\___|
-//
-//                                                  F i l e
-//
-//    This program is free software; you can redistribute it and/or modify
-//    it under the terms of the GNU General Public License as published by
-//    the Free Software Foundation; either version 2 of the License, or
-//    (at your option) any later version.
-//
-//////////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////////
-
-#ifndef COMPILERLOCATORMSVC_H
-#define COMPILERLOCATORMSVC_H
-
-#include <CompilerLocatorMSVCBase.h>
-
-/// Visual Studio 2010 (VC10)
-class CompilerLocatorMSVC10 : public CompilerLocatorMSVCBase
-{
-public:
-    CompilerLocatorMSVC10();
-    virtual ~CompilerLocatorMSVC10();
-
-public:
-    virtual bool Locate();
-};
-
-#endif // COMPILERLOCATORMSVC_H
