+++ Plugin/plugin_sdk.project
@@ -547,8 +547,8 @@
     <File Name="CompilerLocatorGCC.cpp"/>
     <File Name="CompilerLocatorCrossGCC.h"/>
     <File Name="CompilerLocatorCrossGCC.cpp"/>
-    <File Name="CompilerLocatorMSVCBase.h"/>
-    <File Name="CompilerLocatorMSVCBase.cpp"/>
+    <File Name="CompilerLocatorMSVC.h"/>
+    <File Name="CompilerLocatorMSVC.cpp"/>
     <File Name="CompilerLocatorCLANG.h"/>
     <File Name="CompilerLocatorCLANG.cpp"/>
     <File Name="CompilerLocatorCygwin.h"/>
